% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cor_to_cov.R
\name{convert_cor_to_cov}
\alias{convert_cor_to_cov}
\title{Convert correlation matrix to covariance matrix}
\usage{
convert_cor_to_cov(cormat, sd)
}
\arguments{
\item{cormat}{A correlation matrix.}

\item{sd}{A sd vector.}
}
\value{
A covariance matrix.
}
\description{
Convert correlation matrix to covariance matrix
}
\examples{
R <- matrix(c(1, 0.5, 0.5, 1), nrow = 2)
SD <- c(2, 3)
V <- convert_cor_to_cov(R, SD)
}
