% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Sum of the diagonal elements of a square matrix}
\usage{
tr(x)
}
\arguments{
\item{x}{A square matrix}
}
\value{
Sum of the diagonal elements of a square matrix.
}
\description{
Sum of the diagonal elements of a square matrix
}
\examples{
x <- matrix(c(1:9), 3, 3)
tr(x)
}
